+++ projects/ngx-countries-dropdown/src/lib/components/country-list/country-list.component.ts
@@ -1,82 +1,108 @@
 import {
+  ChangeDetectionStrategy,
   Component,
   EventEmitter,
-  HostListener,
-  Input,
   OnInit,
   Output,
+  computed,
+  input,
+  model,
+  signal,
 } from '@angular/core';
 import { IConfig, ICountry } from '../../models';
-import { CountryService } from '../../services/country.service';
+import {
+  getAllowedCountries,
+  getCountriesBasedOnSearch,
+  getFilteredCountries,
+  getPreferredCountries,
+} from '../../helpers/country.helper';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'lib-country-list',
   templateUrl: './country-list.component.html',
   styleUrls: ['./country-list.component.scss'],
+  imports: [FormsModule],
+  standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  host: {
+    '(document:click)': 'onDocumentClick()',
+  },
 })
 export class CountryListComponent implements OnInit {
-  searchText = '';
-  countryList: ICountry[] = [];
-  preferredCountryList: ICountry[] = [];
-  countriesExpectBlocked: ICountry[] = [];
-  filteredCountries: ICountry[] = [];
-  selectedCountry: ICountry = {};
-  displayList = false;
-  @Input() selectedCountryCode: string = '';
-  @Input() preferredCountryCodes: string[] = [];
-  @Input() blockedCountryCodes: string[] = [];
-  @Input() allowedCountryCodes: string[] = [];
-  @Input() selectedCountryConfig: IConfig = {};
-  @Input() countryListConfig: IConfig = {};
-  @Output() onCountryChange = new EventEmitter();
-  constructor(private service: CountryService) {}
+  readonly searchText = model('');
+
+  readonly standardCountries = computed(() =>
+    getFilteredCountries(
+      this.countriesExpectBlocked(),
+      this.preferredCountryCodes()
+    )
+  );
+
+  readonly filteredCountries = computed(() =>
+    getCountriesBasedOnSearch(this.standardCountries(), this.searchText())
+  );
+
+  readonly selectedCountry = signal<ICountry>({});
+
+  readonly displayList = signal(false);
+
+  readonly selectedCountryCode = input('');
+
+  readonly preferredCountryCodes = input<string[]>([]);
+
+  readonly allowedCountryCodes = input<string[]>([]);
+
+  readonly blockedCountryCodes = input<string[]>([]);
+
+  readonly selectedCountryConfig = input<IConfig>({});
+
+  readonly countryListConfig = input<IConfig>({});
+
+  readonly countryList = computed(() =>
+    getAllowedCountries(this.allowedCountryCodes())
+  );
+
+  readonly countriesExpectBlocked = computed(() =>
+    getFilteredCountries(this.countryList(), this.blockedCountryCodes())
+  );
+
+  readonly preferredCountryList = computed(() => {
+    const result = getPreferredCountries(
+      this.countriesExpectBlocked(),
+      this.preferredCountryCodes()
+    );
+
+    return getCountriesBasedOnSearch(result, this.searchText());
+  });
+
+  @Output() onCountryChange = new EventEmitter<ICountry>();
 
   ngOnInit(): void {
-    this.countryList = this.service.getAllAllowedCountries(
-      this.allowedCountryCodes
-    );
-    this.countriesExpectBlocked = this.service.getFilteredCountries(
-      this.countryList,
-      this.blockedCountryCodes
-    );
-    this.preferredCountryList = this.service.setPreferredCountries(
-      this.countriesExpectBlocked,
-      this.preferredCountryCodes
-    );
-    this.filteredCountries = this.service.getStandardCountries();
-    if (this.selectedCountryCode) {
-      const country = this.countriesExpectBlocked.find(
-        (x) => x.code === this.selectedCountryCode.toUpperCase()
+    const selectedCountryCode = this.selectedCountryCode();
+    if (selectedCountryCode) {
+      const country = this.countriesExpectBlocked().find(
+        x => x.code === selectedCountryCode.toUpperCase()
       );
-      if(country){
-        this.selectedCountry= country;
-        this.onCountryChange.emit(this.selectedCountry);
+
+      if (country) {
+        this.selectedCountry.set(country);
+        this.onCountryChange.emit(this.selectedCountry());
       }
     }
   }
 
-  changeCountry(country: ICountry) {
-    this.selectedCountry = country;
-    this.displayList = false;
-    this.onCountryChange.emit(this.selectedCountry);
+  changeCountry(country: ICountry): void {
+    this.selectedCountry.set(country);
+    this.displayList.set(false);
+    this.onCountryChange.emit(this.selectedCountry());
   }
 
-  toggleList() {
-    this.displayList = !this.displayList;
-  }
-  search() {
-    this.preferredCountryList = this.service.getCountriesBasedOnSearch(
-      this.service.getPreferredCountries(),
-      this.searchText
-    );
-    this.filteredCountries = this.service.getCountriesBasedOnSearch(
-      this.service.getStandardCountries(),
-      this.searchText
-    );
+  toggleList(): void {
+    this.displayList.update(isDisplayed => !isDisplayed);
   }
 
-  @HostListener('document:click', ['$event'])
-  onDocumentClick(event: any) {
-    this.displayList = false;
+  onDocumentClick(): void {
+    this.displayList.set(false);
   }
 }