+++ projects/ngx-countries-dropdown/src/lib/ngx-countries-dropdown.module.ts
@@ -1,20 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CountryListComponent } from './components/country-list/country-list.component';
-import { FormsModule } from '@angular/forms';
-import { BrowserModule } from '@angular/platform-browser';
-
-
 
 @NgModule({
-  declarations: [
-    CountryListComponent
-  ],
-  imports: [
-    BrowserModule,
-    FormsModule
-  ],
-  exports: [
-    CountryListComponent
-  ]
+  imports: [CountryListComponent],
+  exports: [CountryListComponent],
 })
-export class NgxCountriesDropdownModule { }
+export class NgxCountriesDropdownModule {}