+++ projects/ngx-countries-dropdown/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
 ## V1.0.3 (06-11-2023)
-- Property to select default dropdown has been added. Consumer need to pass country to Input property 'selectedCountryCode', which accepts string type
 
+- Property to select default dropdown has been added. Consumer need to pass country to Input property 'selectedCountryCode', which accepts string type